+++ .prettierrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  arrowParens: 'avoid',
+  printWidth: 120,
+  semi: false,
+  singleQuote: true,
+  trailingComma: 'none'
+}