+++ src/browser.ts
@@ -1,30 +1,25 @@
-import * as fs from 'fs'
 import * as path from 'path'
 import * as puppeteer from 'puppeteer-core'
 import * as vscode from 'vscode'
 
 import { Buttons } from './buttons'
+import { TreeviewProvider } from './treeview'
 import { WhichChrome } from './whichChrome'
+import { Entry } from './interfaces'
+import { HTTPResponse } from 'puppeteer-core'
 
-const seekMsg = 'Seeking backward/forward function is only work for Youtube videos'
-
+const SEEK_MSG = 'Seeking backward/forward function is only work for Youtube videos. 💡'
+const STATE_MSG = 'Please select the tab/page that either in playing or paused. 💡'
 export class Browser {
   public static activeBrowser: Browser | undefined
   public static cssPath: string
   public static jsPath: string
-  public static launched: boolean = false
-  public static uiHtmlPath: string
-  public static playButtonCss = {
-    soundcloud: '.playControl',
-    spotify: '.control-button--circled',
-    youtube: '.ytp-play-button',
-    ytmusic: '#play-pause-button'
-  }
+  public static launched = false
   private buttons: Buttons
   private currentBrowser: puppeteer.Browser
   private incognitoContext: puppeteer.BrowserContext
-  private pages: puppeteer.Page[] | undefined
-  private selectedMusicPageBrand: string | undefined
+  private pagesStatus: Entry[]
+  private selectedMusicBrand: string | undefined
   private selectedPage: puppeteer.Page | undefined
 
   public static launch(buttons: Buttons, context: vscode.ExtensionContext) {
@@ -43,26 +38,29 @@ export class Browser {
 
       let cPath = vscode.workspace.getConfiguration().get('lmptm.browserPath')
       if (!cPath) cPath = WhichChrome.getPaths().Chrome || WhichChrome.getPaths().Chromium
+
       if (!cPath) {
-        vscode.window.showInformationMessage('Missing Browser! 🤔')
+        vscode.window.showInformationMessage('No Chromium or Chrome browser found. 🤔')
         return
       }
 
-      const links: any = vscode.workspace.getConfiguration().get('lmptm.startPages')
-      if(links.length) {
+      const links: string[] | undefined = vscode.workspace.getConfiguration().get('lmptm.startPages')
+      if (links && links.length) {
         let invalid = false
         links.forEach((e: string) => {
           try { new URL(e) } catch (err) {
             invalid = true
             return
-          }})
-        if (invalid){
-          vscode.window.showErrorMessage('You may have an invalid url on startPages setting! 🤔')
+          }
+        })
+        if (invalid) {
+          vscode.window.showErrorMessage('You may have an invalid url on startPages setting. 🤔')
           return
         }
       }
 
       Browser.launched = true
+      // console.log('###########################################')
       puppeteer.launch({
         args,
         defaultViewport: null,
@@ -71,17 +69,16 @@ export class Browser {
         ignoreDefaultArgs: iArgs
       }).then(async (browser: puppeteer.Browser) => {
         buttons.setStatusButtonText('Running $(browser)')
-        Browser.cssPath = path.join(context.extensionPath, 'dist', 'scripts', 'style.css')
-        Browser.jsPath = path.join(context.extensionPath, 'dist', 'scripts', 'script.js')
-        Browser.uiHtmlPath = fs.readFileSync(path.join(context.extensionPath, 'dist', 'scripts', 'ui.html'), 'utf8')
+        Browser.cssPath = path.join(context.extensionPath, 'dist', 'inject', 'style.css')
+        Browser.jsPath = path.join(context.extensionPath, 'dist', 'inject', 'script.js')
         const defaultPages = await browser.pages()
         defaultPages[0].close() // evaluateOnNewDocument won't on this page
-        Browser.activeBrowser = new Browser(browser, buttons, await browser.createIncognitoBrowserContext())
-        Browser.launched = false
-
-      }, error => {
+        const b = new Browser(browser, buttons, await browser.createIncognitoBrowserContext())
+        Browser.activeBrowser = b
+        TreeviewProvider.refresh()
+      }, (error: { message: string }) => {
         vscode.window.showErrorMessage(error.message)
-        vscode.window.showInformationMessage('Missing Chrome? 🤔')
+        vscode.window.showInformationMessage('Browser launch failed. 😲')
         Browser.launched = false
       })
     }
@@ -90,59 +87,40 @@ export class Browser {
   constructor(browser: puppeteer.Browser, buttons: Buttons, incognitoContext: puppeteer.BrowserContext) {
     this.buttons = buttons
     this.currentBrowser = browser
-    this.pages = undefined
-    this.selectedPage = undefined
+    this.pagesStatus = []
     this.incognitoContext = incognitoContext
-    this.currentBrowser.on('targetcreated', target => this.update('page_created', target))
-    this.currentBrowser.on('targetchanged', target => this.update('page_changed', target))
-    // this.currentBrowser.on('targetdestroyed', target => this.update('page_destroyed',target))
+    this.currentBrowser.on('targetcreated', async (target: puppeteer.Target) => this.update('page_created', await target.page()))
+    this.currentBrowser.on('targetchanged', async (target: puppeteer.Target) => this.update('page_changed', await target.page()))
+    // this.currentBrowser.on('targetdestroyed', target => this.update('page_destroyed', target))
     this.currentBrowser.on('disconnected', () => {
       this.buttons.setStatusButtonText('Launch $(rocket)')
       Browser.activeBrowser = undefined
-      this.buttons.dipslayPlayback(false)
+      this.buttons.displayPlayback(false)
+      TreeviewProvider.refresh()
+      Browser.launched = false
+      // console.debug('CLOSE')
     })
+    // this.currentBrowser.process().once('close', () => console.debug('CLOSE!!!!!!!!'))
     this.launchPages()
   }
 
-  play() {
+  // Toggle
+  async playPause() {
     if (!this.selectedPage) return
-    switch (this.selectedMusicPageBrand) {
-      case 'soundcloud':
-      case 'ytmusic':
+    const { state } = await this.getPlaybackState(this.selectedPage)
+    switch (this.selectedMusicBrand) {
+      case 'soundcloud': case 'spotify': case 'ytmusic':
         this.selectedPage.keyboard.press('Space')
-        break
-      case 'spotify':
-        // @ts-ignore
-        this.selectedPage.evaluate(() => spotifyAction('play'))
         break
       case 'youtube':
         this.selectedPage.keyboard.press('k')
-        break
     }
-    this.buttons.setPlayButton('pause')
-  }
-
-  pause() {
-    if (!this.selectedPage) return
-    switch (this.selectedMusicPageBrand) {
-      case 'soundcloud':
-      case 'ytmusic':
-        this.selectedPage.keyboard.press('Space')
-        break
-      case 'spotify':
-        // @ts-ignore
-        this.selectedPage.evaluate(() => spotifyAction('pause'))
-        break
-      case 'youtube':
-        this.selectedPage.keyboard.press('k')
-        break
-    }
-    this.buttons.setPlayButton('play')
+    this.buttons.setPlayButtonLabel(state)
   }
 
   async skip() {
     if (!this.selectedPage) return
-    switch (this.selectedMusicPageBrand) {
+    switch (this.selectedMusicBrand) {
       case 'soundcloud':
         await this.selectedPage.keyboard.down('ShiftLeft')
         await this.selectedPage.keyboard.press('ArrowRight')
@@ -159,14 +137,12 @@ export class Browser {
         break
       case 'ytmusic':
         await this.selectedPage.keyboard.press('j')
-        break
     }
-    this.changeEventCheck()
   }
 
   async back() {
     if (!this.selectedPage) return
-    switch (this.selectedMusicPageBrand) {
+    switch (this.selectedMusicBrand) {
       case 'soundcloud':
         await this.selectedPage.keyboard.down('ShiftLeft')
         await this.selectedPage.keyboard.press('ArrowLeft')
@@ -181,37 +157,26 @@ export class Browser {
         break
       case 'ytmusic':
         await this.selectedPage.keyboard.press('j')
-        break
     }
-    this.changeEventCheck()
   }
 
   async forward() {
     if (!this.selectedPage) return
-    switch (this.selectedMusicPageBrand) {
+    switch (this.selectedMusicBrand) {
       case 'youtube':
         await this.selectedPage.keyboard.press('ArrowRight')
         break
-      default: { vscode.window.showInformationMessage(seekMsg) }
+      default: vscode.window.showInformationMessage(SEEK_MSG)
     }
-    this.changeEventCheck()
   }
 
   async backward() {
     if (!this.selectedPage) return
-    switch (this.selectedMusicPageBrand) {
+    switch (this.selectedMusicBrand) {
       case 'youtube':
         await this.selectedPage.keyboard.press('ArrowLeft')
         break
-      default: { vscode.window.showInformationMessage(seekMsg) }
-    }
-    this.changeEventCheck()
-  }
-
-  async toggle() {
-    if (this.selectedPage) {
-      const pStt = await this.getPlayingStatus(this.selectedPage)
-      pStt.status === 'play' ? this.pause() : this.play()
+      default: vscode.window.showInformationMessage(SEEK_MSG)
     }
   }
 
@@ -219,18 +184,35 @@ export class Browser {
     return this.selectedPage?.title()
   }
 
+  getPagesStatus() {
+    return this.pagesStatus
+  }
+
+  async pickTab(index: number) {
+    if (!this.pagesStatus) return
+    await this.tabOrderUpdate()
+    const { page, state } = this.pagesStatus[index]
+    if (state === 'none') {
+      vscode.window.showInformationMessage(STATE_MSG)
+      return
+    }
+    this.update('page_selected:tab', page)
+  }
+
+  // ↓↓↓↓ Private methods ↓↓↓↓
+
   private async launchPages() {
-    const links: any  = vscode.workspace.getConfiguration().get('lmptm.startPages')
-    if (links.length) {
-      const p: any = []
+    const links: string[] | undefined = vscode.workspace.getConfiguration().get('lmptm.startPages')
+    if (links && links.length) {
+      const p: Promise<HTTPResponse>[] = []
       links.forEach(async (e: string) => {
         const pg = await this.newPage()
         await pg.setDefaultNavigationTimeout(0)
-        p.push(pg.goto(e))
+        await pg.goto(e)
       })
-      await Promise.all(p)
+      await Promise.all(p) // need to wait?
     }
-    this.pages = await this.currentBrowser.pages()
+    TreeviewProvider.refresh()
   }
 
   private async newPage() {
@@ -239,159 +221,44 @@ export class Browser {
     else return this.currentBrowser.newPage()
   }
 
-  // The button doesn't show up on the 1st launch
-  private injectHtml(page: puppeteer.Page) {
-    page.evaluate(uiHtmlPath => {
-      do {
-        // @ts-ignore
-        if (!window['injected']) {
-          const div = document.createElement('div')
-          div.innerHTML = uiHtmlPath
-          document.getElementsByTagName('body')[0].appendChild(div)
-          // @ts-ignore
-          window['injected'] = true
-        }
-      } while (!document.getElementsByTagName('body')[0])
-    }, Browser.uiHtmlPath)
-  }
-
-  private addScripts(page: puppeteer.Page) {
-    page.addStyleTag({ path: Browser.cssPath })
-    page.addScriptTag({ path: Browser.jsPath })
-  }
-
-  private async setupPageWatcher(page: puppeteer.Page) {
-    page.evaluateOnNewDocument(uiHtmlPath => {
-      window.onload = () => {
-        // @ts-ignore
-        if (!window['injected']) {
-          const div = document.createElement('div')
-          div.innerHTML = uiHtmlPath
-          document.getElementsByTagName('body')[0].appendChild(div)
-          // @ts-ignore
-          window['injected'] = true
-        }
-      }
-    }, Browser.uiHtmlPath)
-
-    page.removeAllListeners('close')
-    page.on('close', async () => {
-      await new Promise(resolve => setTimeout(() => resolve(), 1000))
-      if (Browser.activeBrowser) this.update('page_closed', page.target())
-    })
-
-    // @ts-ignore
-    if (!page._pageBindings.has('pageSelected')) {
-      page.exposeFunction('pageSelected', async e => {
-        this.update('pageSelected', page.target())
-        if (page !== this.selectedPage) {
-          if (this.selectedPage) this.resetButton()
-          this.selectedPage = page
-          this.selectedMusicPageBrand = e.brand
-          this.setupMusicPage()
-          this.buttons.dipslayPlayback(true)
-          this.buttons.setStatusButtonText(await this.selectedPage.title())
-          this.changeEventCheck()
-        }
-      })
-    }
-  }
-
-  private setupMusicPage() {
-    const page = this.selectedPage
-    if (!page) return
-    const brand = this.selectedMusicPageBrand
-    if (!brand) return
-
-    // @ts-ignore
-    if (!page._pageBindings.has('onPlayingChangeEvent')) {
-      page.exposeFunction('onPlayingChangeEvent', () => {
-        this.update('play_event', page.target())
-      })
-    }
-
-    page.evaluate(playButtonCss => {
-      const target = document.querySelector(playButtonCss)
-      // @ts-ignore
-      const observer = new MutationObserver(() => onPlayingChangeEvent())
-      observer.observe(target, { attributes: true })
-      // @ts-ignore
-    }, Browser.playButtonCss[brand])
-
-    if (brand === 'spotify') {
-      page.evaluate(() => {
-        const id = setInterval(() => {
-          if (document.querySelectorAll('.now-playing .cover-art-image')[0]) {
-            const target = document.querySelectorAll('.now-playing .cover-art-image')[0]
-            // @ts-ignore
-            const observer = new MutationObserver(() => onPlayingChangeEvent())
-            observer.observe(target, { attributes: true })
-            clearInterval(id)
-          }
-        }, 3000)
-      })
-    }
-  }
-
-  private async updatePages() {
-    this.pages = await this.currentBrowser.pages()
-    return this.pages[0]
-  }
-
-  private resetButton() {
+  private resetFloatButton() {
     // @ts-ignore
     this.selectedPage?.evaluate(() => reset())
   }
 
-  private async changeEventCheck() {
-    if (!this.selectedPage) return
-    const pStatus = await this.getPlayingStatus(this.selectedPage)
-    if (pStatus.brand !== 'other') {
-      this.selectedMusicPageBrand = pStatus.brand
-      this.buttons.setPlayButton(pStatus.status)
-      if(this.selectedPage)
-        this.buttons.setStatusButtonText(await this.selectedPage.title())
-    } else {
-      if (this.selectedPage.url().includes('www.youtube.com')) this.resetButton() // See line 400
-      this.buttons.setStatusButtonText('Running $(browser)')
-      this.buttons.dipslayPlayback(false)
-      this.selectedPage = undefined
-      this.selectedMusicPageBrand = undefined
-    }
+  private clickFloatButton(thePage: puppeteer.Page) {
+    // @ts-ignore
+    thePage.evaluate(() => click())
   }
 
-  private async getPlayingStatus(page: puppeteer.Page) {
+  private async tabOrderUpdate() {
+    // Puppeteer doesn't keep track pages' order?
+    const pages = await this.currentBrowser.pages()
+    const newPagesStatus = []
+    for (const [i, p] of pages.entries()) {
+      for (const e of this.pagesStatus) {
+        if (p !== e.page) continue
+        e.index = i
+        newPagesStatus.push(e)
+      }
+    }
+    this.pagesStatus = newPagesStatus
+    TreeviewProvider.refresh()
+  }
+
+  // Get from saved
+  private getPlaybackState(page: puppeteer.Page) {
+    for (const e of this.pagesStatus)
+      if (e.page === page) return e
+    // when not found
+    return this._getPlaybackState(page)
+  }
+
+  // Need this?
+  private async _getPlaybackState(page: puppeteer.Page) {
     const pageBrand = this.musicBrandCheck(page.url())
-
-    if (pageBrand === 'other' || !this.selectedPage) return { brand: pageBrand, status: '' }
-
-    else if (pageBrand === 'soundcloud') {
-      const element = await this.selectedPage.$(Browser.playButtonCss.soundcloud)
-      const text = await this.selectedPage.evaluate(element => element.getAttribute('title'), element)
-      const stt = text.includes('Play') ? 'play' : 'pause'
-      return { brand: pageBrand, status: stt }
-
-    } else if (pageBrand === 'spotify') {
-      const element = await this.selectedPage.$(Browser.playButtonCss.spotify)
-      const text = await this.selectedPage.evaluate(element => element.getAttribute('title'), element)
-      const stt = text.includes('Play') ? 'play' : 'pause'
-      return { brand: pageBrand, status: stt }
-
-    } else if (pageBrand === 'youtube') {
-      const element = await this.selectedPage.$(Browser.playButtonCss.youtube)
-      const text = await this.selectedPage.evaluate(element => element.getAttribute('aria-label'), element)
-      if (!text) return { brand: pageBrand, status: 'play' } // When replay
-      const stt = text.includes('Play') ? 'play' : 'pause'
-      return { brand: pageBrand, status: stt }
-
-    } else if (pageBrand === 'ytmusic') {
-      const element = await this.selectedPage.$(Browser.playButtonCss.ytmusic)
-      const text = await this.selectedPage.evaluate(element => element.getAttribute('aria-label'), element)
-      if (!text) return { brand: pageBrand, status: 'play' }
-      const stt = text.includes('Play') ? 'play' : 'pause'
-      return { brand: pageBrand, status: stt }
-
-    } else return { brand: pageBrand, status: '' }
+    const state = await page.evaluate(() => navigator.mediaSession.playbackState)
+    return { brand: pageBrand, state }
   }
 
   private musicBrandCheck(url: string) {
@@ -402,57 +269,162 @@ export class Browser {
     else return 'other'
   }
 
-  private async closeEventUpdate() {
-    this.buttons.setPlayButton('play')
-    this.buttons.dipslayPlayback(false)
-    this.buttons.setStatusButtonText('Running $(browser)')
-    this.selectedPage = undefined
-    this.selectedMusicPageBrand = undefined
-  }
-
-  private async update(event: string, target: puppeteer.Target) {
-    const page = await target.page()
+  private async update(event: string, page: puppeteer.Page | null) {
+    // console.debug('$$$$$$$$$', event)
     if (!page) return
 
-    if (event === 'page_closed') {
-      if (page === this.selectedPage) this.closeEventUpdate()
-      else this.updatePages()
+    let extra
+    if (event.includes('page_selected')) {
+      [event, extra] = event.split(':')
+      if (!extra) throw new Error('page_selected event needs source - tab|button')
+    }
+    else if (event.includes('playback_change')) {
+      [event, extra] = event.split(':')
+      if (!extra) throw new Error('playback_change event needs state - playing|paused|none')
     }
 
-    else if (event === 'page_created') {
-      if ((this.musicBrandCheck(page.url()) === 'spotify')) {
-        this.setPageBypassCSP(page, 'true')
-        page.goto(page.url())
-      } else this.setPageBypassCSP(page, 'false')
-      await this.setupPageWatcher(page)
-
-      page.on('load', async () => {
-        if (this.musicBrandCheck(page.url()) === 'spotify') await this.checkSpotifyCSP(page)
-        this.injectHtml(page)
-        this.addScripts(page)
-        if (!(this.musicBrandCheck(page.url()) === 'other')) this.setupMusicPage()
-      })
+    switch (event) {
+      case 'page_changed': await this.pageChanged(page); break
+      case 'page_closed': this.pageClosed(page); break
+      case 'page_created': await this.pageCreated(page); break
+      case 'page_selected': await this.pageSelected(page, <string>extra); break
+      case 'playback_changed': await this.playbackChanged(page, <string>extra); break
+      default: vscode.window.showErrorMessage(`Unknown event - ${event}`)
     }
-
-    else if (event === 'page_changed') {
-      await page.waitForNavigation()
-      this.changeEventCheck()
-    }
-
-    else if (event === 'play_event') this.changeEventCheck()
+    TreeviewProvider.refresh()
   }
 
-  private async setPageBypassCSP(page: puppeteer.Page, flag: string) {
-    if (page.url() === 'about:blank') return
-    page.setBypassCSP(flag === 'true')
-    await page.evaluate(theFlag => sessionStorage.setItem('bypassCSP', theFlag), flag)
+  private async pageChanged(page: puppeteer.Page) {
+    await page.waitForNetworkIdle() // this somehow prevents navigation error
+    const pageURL = page.url()
+    const brand = this.musicBrandCheck(pageURL)
+
+    // Spotify needs bypass CSP
+    // won't stick to another site (if go to another URL) after set
+    if (brand === 'spotify') this.spotifyBypassCSP(page)
+
+    const title = await page.title()
+
+    for (const [i, e] of this.pagesStatus.entries()) {
+      if (e.page !== page) continue
+      if (this.selectedPage === page) this.selectedMusicBrand = brand
+      if (brand === 'other') {
+        this.pagesStatus[i].picked = false
+        this.pagesStatus[i].state = 'none'
+        this.closingHelper()
+      }
+      this.pagesStatus[i].brand = brand
+      this.pagesStatus[i].title = title
+    }
   }
 
-  private async checkSpotifyCSP(page: puppeteer.Page) {
-    const cspFlag = await page.evaluate(() => sessionStorage.getItem('bypassCSP'))
-    if (cspFlag === 'true') return
-    await this.setPageBypassCSP(page, 'true')
-    await page.reload()
+  private pageClosed(page: puppeteer.Page) {
+    if (page === this.selectedPage) this.closingHelper()
+    this.pagesStatus.forEach((e, i, arr) => {
+      if (e.page !== page) return
+      arr.splice(i, 1)
+    })
+  }
+
+  private closingHelper() {
+    this.buttons.displayPlayback(false)
+    this.buttons.setStatusButtonText('Running $(browser)')
+    this.selectedPage = undefined
+    this.selectedMusicBrand = undefined
+  }
+
+  private async pageCreated(page: puppeteer.Page) {
+    const pageURL = page.url()
+    const brand = pageURL === 'about:blank' ? 'other' : this.musicBrandCheck(pageURL)
+
+    // Spotify needs bypass CSP
+    if (brand === 'spotify') this.spotifyBypassCSP(page)
+
+    let title = pageURL === 'about:blank' ? pageURL : await page.title()
+    if (title === '') title = 'New Tab'
+
+    const pages = await this.currentBrowser.pages()
+    for (const [index, p] of pages.entries()) {
+      if (p !== page) continue
+      this.pagesStatus.splice(index, 0, { page, brand, index, state: 'none', picked: false, title })
+    }
+
+    page.on('load', async () => {
+      page.addStyleTag({ path: Browser.cssPath })
+      page.addScriptTag({ path: Browser.jsPath })
+    })
+
+    page.removeAllListeners('close')
+    page.on('close', async () => {
+      // console.debug('page on CLOSE')
+      await new Promise<void>(resolve => setTimeout(() => resolve(), 1000))
+      if (Browser.activeBrowser) this.update('page_closed', page)
+    })
+
+    // @ts-ignore
+    if (!page._pageBindings.has('pageSelected'))
+      page.exposeFunction('pageSelected', () => this.update('page_selected:button', page))
+
+    // @ts-ignore
+    if (!page._pageBindings.has('playbackChanged'))
+      page.exposeFunction('playbackChanged', (state: string) => this.update(`playback_changed:${state}`, page))
+
+  }
+
+  private async pageSelected(page: puppeteer.Page, source: string) {
+    this.buttons.displayPlayback(true)
+
+    // Not sure why it can't detect or wait - the error below
+    // rejected promise not handled within 1 second:
+    // Error: Execution context is not available in detached frame "about:blank"
+    // (are you trying to evaluate?)
+
+    if (this.selectedPage) {
+      const { state } = await this.getPlaybackState(this.selectedPage)
+      if (this.selectedPage === page) {
+        await this.playPause()
+        return
+      } else {
+        this.pagesStatus.forEach(e => e.page === this.selectedPage ? e.picked = false : null)
+        if (source === 'button' && state === 'playing') await this.playPause()
+        this.resetFloatButton()
+      }
+    }
+
+    for (const [i, e] of this.pagesStatus.entries()) {
+      if (e.page !== page) continue
+      if (source === 'tab') {
+        this.clickFloatButton(e.page) // this will trigger page_select:button
+        break
+      } else {
+        this.selectedPage = e.page
+        this.pagesStatus[i].picked = true
+        this.selectedMusicBrand = this.musicBrandCheck(page.url())
+        const title = await this.selectedPage.title()
+        this.pagesStatus[i].title = title
+        this.buttons.setStatusButtonText(title)
+      }
+    }
+  }
+
+  private async playbackChanged(page: puppeteer.Page, state: string) {
+    for (const [i, e] of this.pagesStatus.entries()) {
+      if (e.page !== page) continue
+      if (this.selectedPage === page) {
+        this.buttons.setPlayButtonLabel(<MediaSessionPlaybackState>state)
+        this.buttons.setStatusButtonText(await this.selectedPage.title())
+      }
+      this.pagesStatus[i].state = <MediaSessionPlaybackState>state
+      this.pagesStatus[i].title = await page.title()
+    }
+  }
+
+  private async spotifyBypassCSP(page: puppeteer.Page) {
+    page.setBypassCSP(true)
+    // for debugging
+    await page.evaluate(() => sessionStorage.setItem('bypassCSP', 'true'))
+    // this doesn't trigger page_changed again -> no infinity loop
+    page.goto(page.url())
   }
 
   // private async sleep(ms: number = 1000) {