+++ src/buttons.ts
@@ -17,8 +17,8 @@ export class Buttons {
     this.backButton.text = '$(chevron-left)'
     this.statusButton.text = 'Launch $(rocket)'
 
-    this.statusButton.command = 'lmptm.browserlaunch'
-    this.playButton.command = 'lmptm.play'
+    this.statusButton.command = 'lmptm.browserLaunch'
+    this.playButton.command = 'lmptm.playPause'
     this.backButton.command = 'lmptm.back'
     this.skipButton.command = 'lmptm.skip'
 
@@ -28,7 +28,7 @@ export class Buttons {
   setStatusButtonText(text: string) {
     if (text === 'Launch $(rocket)') {
       this.statusButton.text = text
-      this.statusButton.command = 'lmptm.browserlaunch'
+      this.statusButton.command = 'lmptm.browserLaunch'
     } else if (text === 'Running $(browser)') {
       this.statusButton.text = text
       this.statusButton.command = undefined
@@ -41,16 +41,11 @@ export class Buttons {
     }
   }
 
-  setPlayButton(text: string) {
-    if (text === 'play') {
-      this.playButton.text = '$(play)'
-      this.playButton.command = 'lmptm.play'
-    } else {
-      this.playButton.text = '$(primitive-square)'
-      this.playButton.command = 'lmptm.pause'
-    }
+  setPlayButtonLabel(label: MediaSessionPlaybackState) {
+    this.playButton.text = label === 'playing' ? '$(primitive-square)' : '$(play)'
   }
-  dipslayPlayback(flag: boolean) {
+
+  displayPlayback(flag: boolean) {
     if (flag) {
       this.playButton.show()
       this.skipButton.show()