+++ src/extension.ts
@@ -1,19 +1,21 @@
 import { commands, window, ExtensionContext } from 'vscode'
-import { Buttons } from './buttons'
 import { Browser } from './browser'
+import { Buttons } from './buttons'
+import { Entry } from './interfaces'
+import { TreeviewProvider } from './treeview'
 
 export function activate(context: ExtensionContext) {
   const buttons = new Buttons()
+  const fn = ['playPause', 'skip', 'back', 'forward', 'backward'] as const
+  const rc = commands.registerCommand
+  TreeviewProvider.create()
+  const disposables = fn.map(n => rc(`lmptm.${n}`, () => Browser.activeBrowser?.[n]()))
+  context.subscriptions.concat(disposables)
   context.subscriptions.concat([
-    commands.registerCommand('lmptm.browserlaunch', () => Browser.launch(buttons, context)),
-    commands.registerCommand('lmptm.play', () => Browser.activeBrowser?.play()),
-    commands.registerCommand('lmptm.pause', () => Browser.activeBrowser?.pause()),
-    commands.registerCommand('lmptm.skip', () => Browser.activeBrowser?.skip()),
-    commands.registerCommand('lmptm.back', () => Browser.activeBrowser?.back()),
-    commands.registerCommand('lmptm.forward', () => Browser.activeBrowser?.forward()),
-    commands.registerCommand('lmptm.backward', () => Browser.activeBrowser?.backward()),
-    commands.registerCommand('lmptm.toggle', () => Browser.activeBrowser?.toggle()),
-    commands.registerCommand('lmptm.showTitle', showTitle)
+    rc('lmptm.browserLaunch', () => Browser.launch(buttons, context)),
+    rc('lmptm.tvRefresh', () => TreeviewProvider.refresh()),
+    rc('lmptm.showTitle', showTitle),
+    rc('lmptm.click', selection => click(selection))
   ])
 }
 
@@ -23,4 +25,8 @@ async function showTitle() {
   else window.showErrorMessage('Failed to retrieve title')
 }
 
-export function deactivate() {}
+function click(selection: Entry) {
+  Browser.activeBrowser?.pickTab(selection.index)
+}
+
+// export function deactivate() { }