+++ CHANGELOG.md
@@ -4,10 +4,49 @@ All notable changes to the "LetMePlayTheMusic" extension will be documented in t
 
 Check [Keep a Changelog](http://keepachangelog.com) for recommendations on how to structure this file.
 
+### References
+- https://github.com/microsoft/vscode-generator-code/tree/main/generators/app/templates/ext-command-ts
+- https://github.com/microsoft/vscode-extension-samples
+- https://www.conventionalcommits.org
+
 ### TODO
-- Extention settings scope
-- I18n?
-- Site English version issue
+- Brave setting removal
+- I18n
+- Playwright vs Puppeteer option
+- Seek backward/forward setting option
+- Shelljs vs Execa
+- Support for other sites
+
+## [2.0.0] - December 2021
+- Add new feature - treeview
+- Rewrite/refactor most of the code - 30+ commits
+- New float button style - old vs new in Windows
+  <br><img src='./media/btn1.4.png' width='100'/> <img src='./media/btn2.0.png' width='100'/>
+- Fix Spotify bug due to it's style class's changes
+- Fix site English version issue - observes another DOM element to update playback status
+- Merge play and pause commands into one toggle function
+- Move most of the minor inject action to inject script instead
+- Re-config project's configs like eslint, tsconfig, vscode setting, and webpack (.js -> .ts)
+- Rename directories: icon -> media, scripts -> inject
+- Playback icons is sync with the site's playback icons which was behaved contrarily before
+- Webpack 5.65.0 compiled successfully in 9980 ms
+- 15 files, 1.39MB, 1.63.0
+
+### Note
+
+#### This release mostly has more significant impact on the dev side rather than like a product update.
+
+#### One funny thing is that the picture in README.md accounts for a large part of this extension's size.
+
+#### [MediaSession](https://developer.mozilla.org/en-US/docs/Web/API/MediaSession)
+- Soundcloud does update `playbackState` but only call `set` for from press play
+- Spotify does call `set` to update metadata but playbackState always *none*
+- Youtube and YTmusic update `playbackState` consistent with proxy `set` event
+
+#### Puppeteer
+- Seem Puppeteer doesn't keep track pages/tabs' order
+- If click too quick, this shows up: `Error: Execution context is not available in detached frame "about:blank" (are you trying to evaluate?)`
+  Waiting and try/catch was fruitless
 
 ## [1.4.0] - October 2020
 - Add key shortcuts [#7](https://github.com/lanly-dev/VSCode-LMPTM/issues/7)
@@ -24,7 +63,7 @@ Check [Keep a Changelog](http://keepachangelog.com) for recommendations on how t
 - 11 files, 141.77KB, 1.42.0
 
 ## [1.2.0] - December 2019
-- Add browser execuatble file setting [#3](https://github.com/lanly-dev/VSCode-LMPTM/issues/3)
+- Add browser executable file setting [#3](https://github.com/lanly-dev/VSCode-LMPTM/issues/3)
 - Add Incognito/Private mode setting
 - Add [User Data Directory](https://chromium.googlesource.com/chromium/src/+/master/docs/user_data_dir.md) setting [#4](https://github.com/lanly-dev/VSCode-LMPTM/issues/4)
 - 11 files, 192KB, 1.41.0